requirements.txt
boto3
python-dotenv
pyspark
delta-spark